Shooter motors
Added code to control all 4 motors mounted on the motor panel. All have
been tested and work as expected. Part of the code for the encoder and
deriving values from the encoder such as counts per second has been
added but is not finished. The PID loop for the shooter has also been
set up but the target velocity must be calculated first before we can
get this to work